\name{currencies_list_CoinMarketCap}
\alias{currencies_list_CoinMarketCap}
\docType{data}
\title{
List of Currencies
}
\description{
list of all availble currencies on the coinmaketcap api
}
\format{
  The format is:
 chr [1:32] "AUD" "BRL" "CAD" "CHF" "CLP" "CNY" "CZK" "DKK" "EUR" "GBP" "HKD" "HUF" "IDR" "ILS" "INR" "JPY" "KRW" "MXN" "MYR" "NOK" "NZD" "PHP" "PKR" "PLN" "RUB" "SEK" "SGD" ...
}
\source{
https://coinmarketcap.com/api/
}
\examples{
data(currencies_list_CoinMarketCap)
## maybe str(currencies_list_CoinMarketCap) ; plot(currencies_list_CoinMarketCap) ...
}

